+++ packages/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zmonitor/types
 
+## 5.2.1
+
+### Patch Changes
+
+- 新增版本以及环境控制
+- Updated dependencies
+  - @zmonitor/common@5.2.1
+
 ## 5.2.0
 
 ### Minor Changes